feat(repositories): add a Setup tab with the per-repo client-setup guide (#560)
